refactor: properties components (#820)
* refactor: properties components

* fix: renamed journey url to journey link

* fix: linting issue

* fix: translations issue and updated tests
