+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyLink/EmbedJourneyDialog/EmbedJourneyDialog.stories.tsx
@@ -2,8 +2,8 @@ import { Story, Meta } from '@storybook/react'
 import { useState } from 'react'
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 import { MockedProvider } from '@apollo/client/testing'
-import { journeysAdminConfig } from '../../../../libs/storybook'
-import { publishedJourney } from '../../data'
+import { journeysAdminConfig } from '../../../../../libs/storybook'
+import { publishedJourney } from '../../../data'
 import { EmbedJourneyDialog } from './EmbedJourneyDialog'
 
 const EmbedJourneyDialogStory = {
