+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyLink/SlugDialog/SlugDialog.tsx
@@ -5,8 +5,8 @@ import { useSnackbar } from 'notistack'
 import { Formik, Form, FormikValues, FormikHelpers } from 'formik'
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
 import { useTranslation } from 'react-i18next'
-import { JourneySlugUpdate } from '../../../../../__generated__/JourneySlugUpdate'
-import { Dialog } from '../../../Dialog'
+import { JourneySlugUpdate } from '../../../../../../__generated__/JourneySlugUpdate'
+import { Dialog } from '../../../../Dialog'
 
 export const JOURNEY_SLUG_UPDATE = gql`
   mutation JourneySlugUpdate($id: ID!, $input: JourneyUpdateInput!) {
