+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyLink/index.ts
@@ -1,0 +1,1 @@
+export { JourneyLink } from './JourneyLink'
